+++ src/main/java/com/yanny/ytech/configuration/block/BronzeAnvilBlock.java
@@ -3,7 +3,7 @@ package com.yanny.ytech.configuration.block;
 import com.mojang.serialization.MapCodec;
 import com.yanny.ytech.configuration.*;
 import com.yanny.ytech.configuration.block_entity.BronzeAnvilBlockEntity;
-import com.yanny.ytech.configuration.recipe.RemainingShapedRecipe;
+import com.yanny.ytech.configuration.recipe.WorkspaceCraftingRecipe;
 import com.yanny.ytech.registration.YTechBlocks;
 import com.yanny.ytech.registration.YTechItemTags;
 import net.minecraft.core.BlockPos;
@@ -206,12 +206,17 @@ public class BronzeAnvilBlock extends FallingBlock implements EntityBlock {
     }
 
     public static void registerRecipe(@NotNull RecipeOutput recipeConsumer) {
-        RemainingShapedRecipe.Builder.shaped(RecipeCategory.MISC, YTechBlocks.BRONZE_ANVIL.get())
+        WorkspaceCraftingRecipe.Builder.recipe(YTechBlocks.BRONZE_ANVIL.get())
                 .define('B', YTechItemTags.STORAGE_BLOCKS.get(MaterialType.BRONZE))
-                .define('I', YTechItemTags.INGOTS.get(MaterialType.BRONZE))
-                .pattern("BBB")
-                .pattern(" I ")
-                .pattern("III")
+                .bottomPattern(" B ")
+                .bottomPattern("BBB")
+                .bottomPattern(" B ")
+                .middlePattern("   ")
+                .middlePattern(" B ")
+                .middlePattern("   ")
+                .topPattern("BBB")
+                .topPattern("BBB")
+                .topPattern("BBB")
                 .unlockedBy(Utils.getHasName(), RecipeProvider.has(YTechItemTags.INGOTS.get(MaterialType.BRONZE)))
                 .save(recipeConsumer, Utils.modLoc(YTechBlocks.BRONZE_ANVIL));
     }