+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # YTech Changelog
 
+## []
+
+- Changed mob spawn biomes
+- Fixed blocks not correctly saving state
+- Added in-world crafting using Crafting Workspace
+- Added Shears
+
 ## [0.6.1]
 
 - Fixed item duplication in FirePit and Drying Rack